requirements.txt
BeautifulSoup==3.2.1
Cython==0.24
Jinja2==2.7.2
MarkupSafe==0.18
PAM==0.4.2
Pillow==2.3.0
PyYAML==3.11
Pygments==1.6
Sphinx==1.2.2
Twisted-Core==13.2.0
Twisted-Web==13.2.0
adium-theme-ubuntu==0.3.4
appdirs==1.4.0
apt-xapian-index==0.45
argparse==1.2.1
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
chainer==1.21.0
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
cssselect==1.0.1
dask==0.9.0
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
docutils==0.11
duplicity==0.6.23
filelock==2.0.7
gyp==0.1
h5py==2.6.0
html5lib==0.999
httplib2==0.8
iniparse==0.4
ipython==4.2.0
ipython-genutils==0.1.0
leveldb==0.193
lockfile==0.8
lxml==3.3.3
matplotlib==1.3.1
mecab-python==0.996
networkx==1.11
nose==1.3.1
numpy==1.12.0
oauthlib==0.6.1
oneconf==0.3.7.14.04.1
packaging==16.8
pandas==0.13.1
pathlib2==2.1.0
pexpect==3.1
pickleshare==0.7.2
piston-mini-client==0.7.5
protobuf==2.6.1
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pyparsing==2.1.10
pyserial==2.6
pysmbc==1.0.14.1
pysqlite==1.0.1
python-apt==0.9.3.5ubuntu2
python-dateutil==2.6.0
python-debian==0.1.21-nmu2ubuntu2
python-gflags==3.0.5
pytz==2016.10
pyxdg==0.25
reportlab==3.0
requests==2.2.1
roman==2.0.0
scikit-image==0.9.3
scipy==0.13.3
selenium==3.0.2
sessioninstaller==0.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
software-center-aptd-plugins==0.0.0
ssh-import-id==3.21
system-service==0.1.6
toolz==0.8.0
tornado==4.3
traitlets==4.2.1
unity-lens-photos==1.0
urlgrabber==3.9.1
urllib3==1.7.1
wget==3.2
wheel==0.24.0
wsgiref==0.1.2
xdiagnose==3.6.3build2
yum-metadata-parser==1.1.4
zope.interface==4.0.5
